File Restructuring and File Python Course Section Inclusion for Black Jack Programme
