+++ Python_Youtube/Fundamentals_of_Python/N_VariableScope.py
