+++ Python_Youtube/Fundamentals_of_Python/T_LamdbaFunctions.py
